test cmpc

+++ chance_mpc/objective.cc
@@ -9,11 +9,11 @@ using namespace Eigen;
 namespace cmpc {
 obj::obj() {}
 
-float obj::step_cost(const Eigen::VectorXf& coeff, const float& t_sample,
-                     const boost::math::cubic_b_spline<float>* ref_path_x,
-                     const boost::math::cubic_b_spline<float>* ref_path_y,
-                     const float& length, const double* x) {
-  VectorXf terms(6);
+double obj::step_cost(const Eigen::VectorXd& coeff, const double& t_sample,
+                      const boost::math::cubic_b_spline<float>* ref_path_x,
+                      const boost::math::cubic_b_spline<float>* ref_path_y,
+                      const float& length, const double* x) {
+  VectorXd terms(6);
   obj::error(x, ref_path_x, ref_path_y, terms(0), terms(1));
 
   terms(2) = x[6] * t_sample;  // progress
@@ -28,7 +28,7 @@ float obj::step_cost(const Eigen::VectorXf& coeff, const float& t_sample,
 void obj::error(const double* x,
                 const boost::math::cubic_b_spline<float>* ref_path_x,
                 const boost::math::cubic_b_spline<float>* ref_path_y,
-                float& e_contour, float& e_lag) {
+                double& e_contour, double& e_lag) {
   float s = x[7];
   float x_ref = (*ref_path_x)(s);
   float y_ref = (*ref_path_y)(s);
