+++ chance_mpc/objective.h
@@ -9,14 +9,14 @@ class obj {
   obj();
 
   /// step cost
-  static float step_cost(const Eigen::VectorXf& coeff, const float& t_sample,
-                         const boost::math::cubic_b_spline<float>* ref_path_x,
-                         const boost::math::cubic_b_spline<float>* ref_path_y,
-                         const float& length, const double* x);
+  static double step_cost(const Eigen::VectorXd& coeff, const double& t_sample,
+                          const boost::math::cubic_b_spline<float>* ref_path_x,
+                          const boost::math::cubic_b_spline<float>* ref_path_y,
+                          const float& length, const double* x);
   /// Contour & lag error
   static void error(const double* x,
                     const boost::math::cubic_b_spline<float>* ref_path_x,
                     const boost::math::cubic_b_spline<float>* ref_path_y,
-                    float& e_contour, float& e_lag);
+                    double& e_contour, double& e_lag);
 };
 }  // namespace cmpc
